requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
duckduckgo_search==8.0.4
ftfy==6.3.1
google_search_results==2.4.2
googlesearch-python==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==5.4.0
mysql-connector-python==9.3.0
openai==1.87.0
primp==0.15.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
